undo unneeded moves and changes for many functionality tests
