+++ libensemble/tests/unit_tests/mpich-only_test_api.py
@@ -67,7 +67,7 @@ def test_bad_func_loads():
 
 
 @pytest.mark.extra
-def test_full_workflow_from_specs():
+def test_full_workflow():
     """Test initializing a workflow via Specs and Ensemble.run()"""
     from libensemble.api import Ensemble
     from libensemble.specs import ExitCriteria, GenSpecs, LibeSpecs, SimSpecs
@@ -92,37 +92,8 @@ def test_full_workflow_from_specs():
         assert len(ens.H) >= 101
 
 
-@pytest.mark.extra
-def test_full_workflow_from_files():
-    """Test initializing a workflow via files and Ensemble.run()"""
-    from libensemble.api import Ensemble
-
-    sampling = Ensemble(libE_specs={"comms": "local", "nworkers": 4})
-    sampling.from_json("./simdir/1d_sampling.json")
-    sampling.from_toml("./simdir/1d_sampling.toml")
-    sampling.from_yaml("./simdir/1d_sampling.yaml")
-
-    sampling.gen_specs["user"].update(
-        {
-            "lb": np.array([-3]),
-            "ub": np.array([3]),
-        }
-    )
-
-    sampling.add_random_streams()
-
-    # Perform the run
-    sampling.run()
-
-    if sampling.is_manager:
-        assert len(sampling.H) >= 201
-        print("\nlibEnsemble with random sampling has generated enough points")
-        sampling.save_output(__file__)
-
-
 if __name__ == "__main__":
     test_ensemble_init()
     test_from_files()
     test_bad_func_loads()
-    test_full_workflow_from_specs()
-    test_full_workflow_from_files()
+    test_full_workflow()
