--- libensemble/tests/unit_tests/simdir/1d_sampling.json
@@ -1,34 +1,0 @@
-{
-  "libE_specs": {
-    "save_every_k_gens": 100,
-    "safe_mode": false,
-    "nworkers": 4,
-    "comms": "local"
-  },
-  "exit_criteria": {
-    "gen_max": 201
-  },
-  "sim_specs": {
-    "sim_f": "libensemble.sim_funcs.one_d_func.one_d_example",
-    "inputs": [
-      "x"
-    ],
-    "out": {
-      "f": {
-        "type": "float"
-      }
-    }
-  },
-  "gen_specs": {
-    "gen_f": "libensemble.gen_funcs.sampling.latin_hypercube_sample",
-    "out": {
-      "x": {
-        "type": "float",
-        "size": 1
-      }
-    },
-    "user": {
-      "gen_batch_size": 200
-    }
-  }
-}
