--- libensemble/tests/unit_tests/simdir/1d_sampling.toml
@@ -1,24 +1,0 @@
-[libE_specs]
-  save_every_k_gens = 100
-  safe_mode = false
-  nworkers = 4
-  comms = "local"
-
-[exit_criteria]
-  gen_max = 201
-
-[sim_specs]
-  sim_f = "libensemble.sim_funcs.one_d_func.one_d_example"
-  inputs = ["x"]
-  [sim_specs.out]
-    [sim_specs.out.f]
-      type = "float"
-
-[gen_specs]
-  gen_f = "libensemble.gen_funcs.sampling.latin_hypercube_sample"
-  [gen_specs.out]
-    [gen_specs.out.x]
-      type = "float"
-      size = 1
-  [gen_specs.user]
-    gen_batch_size = 200
